requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
langchain==0.2.5
langchain-community==0.2.5
langchain-openai==0.1.7
langchain-text-splitters==0.2.2
faiss-cpu==1.7.4
sentence-transformers==2.7.0
pypdf==4.2.0
python-dotenv==1.0.1
pydantic==2.7.1
openai==1.30.1
httpx==0.27.2